+++ lib/client/variables.go
@@ -2,6 +2,7 @@ package client
 
 import (
 	"fmt"
+	"os"
 	"sync"
 
 	"github.com/hashicorp/go-tfe"
@@ -10,7 +11,14 @@ import (
 	log "github.com/sirupsen/logrus"
 )
 
-func (c *Client) ProcessAllVariables(cfg *schemas.Config) error {
+type RenderVariablesType string
+
+const (
+	RenderVariablesTypeTFC   RenderVariablesType = "tfc"
+	RenderVariablesTypeLocal RenderVariablesType = "local"
+)
+
+func (c *Client) RenderVariables(cfg *schemas.Config, t RenderVariablesType, dryRun bool) error {
 	variables := schemas.Variables{}
 
 	for _, variable := range cfg.TerraformVariables {
@@ -23,30 +31,25 @@ func (c *Client) ProcessAllVariables(cfg *schemas.Config) error {
 		variables = append(variables, variable)
 	}
 
+	switch t {
+	case RenderVariablesTypeTFC:
+		log.Info("Processing variables and updating their values on TFC")
+		return c.renderVariablesOnTFC(cfg, variables, dryRun)
+	case RenderVariablesTypeLocal:
+		log.Info("Processing variables and updating their values locally")
+		return c.renderVariablesLocally(variables)
+	default:
+		return fmt.Errorf("undefined ProcessVaribleType '%s'", t)
+	}
+}
+
+func (c *Client) renderVariablesOnTFC(cfg *schemas.Config, vars schemas.Variables, dryRun bool) error {
 	w, err := c.getWorkspace(cfg.TFC.Organization, cfg.TFC.Workspace)
 	if err != nil {
 		return fmt.Errorf("terraform cloud: %s", err)
 	}
-
 	log.Debugf("workspace id for %s: %s", w.Name, w.ID)
 
-	return c.processVariables(w, variables)
-}
-
-func (c *Client) isVariableAlreadyProcessed(name string, kind schemas.VariableKind) bool {
-	c.ProcessedVariablesMutex.Lock()
-	defer c.ProcessedVariablesMutex.Unlock()
-	k, exists := c.ProcessedVariables[name]
-
-	if exists && k == kind {
-		return true
-	}
-
-	c.ProcessedVariables[name] = kind
-	return false
-}
-
-func (c *Client) processVariables(w *tfe.Workspace, vars schemas.Variables) error {
 	// Find existing variables on TFC
 	e, err := c.listVariables(w)
 	if err != nil {
@@ -58,10 +61,53 @@ func (c *Client) processVariables(w *tfe.Workspace, vars schemas.Variables) erro
 
 	for _, v := range vars {
 		wg.Add(1)
-		go func(c *Client, w *tfe.Workspace, v *schemas.Variable, e TFEVariables) {
+		go func(v *schemas.Variable) {
 			defer wg.Done()
-			ch <- c.processVariable(w, v, e)
-		}(c, w, v, e)
+			ch <- c.renderVariableOnTFC(w, v, e, dryRun)
+		}(v)
+	}
+
+	go func() {
+		wg.Wait()
+		close(ch)
+	}()
+
+	for err := range ch {
+		if err != nil {
+			return err
+		}
+	}
+
+	if cfg.TFC.PurgeUnmanagedVariables != nil && *cfg.TFC.PurgeUnmanagedVariables {
+		log.Debugf("Looking for unmanaged variables to remove")
+		return c.purgeUnmanagedVariables(vars, e, dryRun)
+	}
+
+	return nil
+}
+
+func (c *Client) renderVariablesLocally(vars schemas.Variables) error {
+	envFile, err := os.OpenFile("./tfcw.env", os.O_TRUNC|os.O_CREATE|os.O_WRONLY, 0644)
+	if err != nil {
+		return err
+	}
+	defer envFile.Close()
+
+	tfFile, err := os.OpenFile("./tfcw.auth.tfvars", os.O_TRUNC|os.O_CREATE|os.O_WRONLY, 0644)
+	if err != nil {
+		return err
+	}
+	defer tfFile.Close()
+
+	ch := make(chan error)
+	wg := sync.WaitGroup{}
+
+	for _, v := range vars {
+		wg.Add(1)
+		go func(v *schemas.Variable) {
+			defer wg.Done()
+			ch <- c.renderVariableLocally(v, envFile, tfFile)
+		}(v)
 	}
 
 	go func() {
@@ -77,7 +123,45 @@ func (c *Client) processVariables(w *tfe.Workspace, vars schemas.Variables) erro
 	return nil
 }
 
-func (c *Client) processVariable(w *tfe.Workspace, v *schemas.Variable, e TFEVariables) error {
+func (c *Client) renderVariableOnTFC(w *tfe.Workspace, v *schemas.Variable, e TFEVariables, dryRun bool) error {
+	c.fetchVariableValue(v)
+	if !dryRun {
+		if _, err := c.setVariableOnTFC(w, v, e); err != nil {
+			return err
+		}
+	}
+
+	logVariable(v, dryRun)
+	return nil
+}
+
+func (c *Client) renderVariableLocally(v *schemas.Variable, envFile, tfFile *os.File) error {
+	c.fetchVariableValue(v)
+	switch v.Kind {
+	case schemas.VariableKindEnvironment:
+		if _, err := envFile.WriteString(fmt.Sprintf("export %s=%s\n", v.Name, *v.Value)); err != nil {
+			return err
+		}
+	case schemas.VariableKindTerraform:
+		s := ""
+		if *v.HCL {
+			s = fmt.Sprintf("%s = %s\n", v.Name, *v.Value)
+		} else {
+			s = fmt.Sprintf("%s = \"%s\"\n", v.Name, *v.Value)
+		}
+
+		if _, err := tfFile.WriteString(s); err != nil {
+			return err
+		}
+	default:
+		return fmt.Errorf("unkown kind '%s' for variable %s", v.Kind, v.Name)
+	}
+
+	logVariable(v, false)
+	return nil
+}
+
+func (c *Client) fetchVariableValue(v *schemas.Variable) error {
 	if c.isVariableAlreadyProcessed(v.Name, v.Kind) {
 		return fmt.Errorf("duplicate variable '%s' (%s)", v.Name, v.Kind)
 	}
@@ -96,14 +180,14 @@ func (c *Client) processVariable(w *tfe.Workspace, v *schemas.Variable, e TFEVar
 	}
 
 	if configuredProviders != 1 {
-		return fmt.Errorf("You can't have more or less than one provider configured per variable. Found %d for '%s'", configuredProviders, v.Name)
+		return fmt.Errorf("you can't have more or less than one provider configured per variable. Found %d for '%s'", configuredProviders, v.Name)
 	}
 
 	// We can map several keys in a single API call
 	if v.Vault != nil && v.Vault.Path != nil {
 		if values, err := c.Vault.GetValues(v.Vault); err == nil {
 			if v.Vault.Key == nil && (v.Vault.Keys == nil || len(*v.Vault.Keys) == 0) {
-				return fmt.Errorf("You either need to set 'key' or 'keys' when using the Vault provider")
+				return fmt.Errorf("you either need to set 'key' or 'keys' when using the Vault provider")
 			}
 
 			if v.Vault.Keys == nil {
@@ -118,13 +202,9 @@ func (c *Client) processVariable(w *tfe.Workspace, v *schemas.Variable, e TFEVar
 				if value, found := values[vaultKey]; found {
 					v.Name = variableName
 					v.Value = &value
-					if _, err := c.setVariable(w, v, e); err != nil {
-						return err
-					}
-					logVariable(v)
-				} else {
-					return fmt.Errorf("key '%s' was not found in secret '%s'", vaultKey, *v.Vault.Path)
+					return nil
 				}
+				return fmt.Errorf("key '%s' was not found in secret '%s'", vaultKey, *v.Vault.Path)
 			}
 		}
 	}
@@ -136,30 +216,37 @@ func (c *Client) processVariable(w *tfe.Workspace, v *schemas.Variable, e TFEVar
 		}
 
 		v.Value = &value
-		if _, err := c.setVariable(w, v, e); err != nil {
-			return err
-		}
-		logVariable(v)
+		return nil
 	}
 
 	if v.Env != nil {
 		value := c.Env.GetValue(v.Env)
 		v.Value = &value
-		if _, err := c.setVariable(w, v, e); err != nil {
-			return err
-		}
-		logVariable(v)
+		return nil
 	}
-	return nil
+
+	return fmt.Errorf("no providers could be used to fetch the variable value")
 }
 
-func logVariable(v *schemas.Variable) error {
-	log.WithFields(log.Fields{
-		"kind":  v.Kind,
-		"name":  v.Name,
-		"value": secureSensitiveString(*v.Value),
-	}).Infof("set!")
+func (c *Client) isVariableAlreadyProcessed(name string, kind schemas.VariableKind) bool {
+	c.ProcessedVariablesMutex.Lock()
+	defer c.ProcessedVariablesMutex.Unlock()
+	k, exists := c.ProcessedVariables[name]
 
+	if exists && k == kind {
+		return true
+	}
+
+	c.ProcessedVariables[name] = kind
+	return false
+}
+
+func logVariable(v *schemas.Variable, dryRun bool) error {
+	if dryRun {
+		log.Infof("[DRY-RUN] Set variable %s - (%s) : %s", v.Name, v.Kind, secureSensitiveString(*v.Value))
+	} else {
+		log.Infof("Set variable %s (%s)", v.Name, v.Kind)
+	}
 	return nil
 }