+++ lib/schemas/tfc.go
@@ -1,6 +1,7 @@
 package schemas
 
 type TFC struct {
-	Organization string `hcl:"organization"`
-	Workspace    string `hcl:"workspace"`
+	Organization            string `hcl:"organization"`
+	Workspace               string `hcl:"workspace"`
+	PurgeUnmanagedVariables *bool  `hcl:"purge-unmanaged-variables"`
 }