+++ CHANGELOG.md
@@ -11,5 +11,8 @@ and this project adheres to [0ver](https://0ver.org).
 
 - Read configuration form HCL (or json) file
 - Fetch sensitive values from 3 providers : `vault`, `s5` and `environment variables`
+- Plan and apply Terraform stacks
+- dry-run feature on render function
+- purge unmanaged variables
 
 [Unreleased]: https://github.com/mvisonneau/tfcw/compare/0.0.0...HEAD